+++ components/zcash_protocol/CHANGELOG.md
@@ -6,3 +6,18 @@ and this library adheres to Rust's notion of
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+The entries below are relative to the `zcash_primitives` crate as of the tag
+`zcash_primitives-0.13.0`.
+
+### Added
+- The following modules have been extracted from `zcash_primitives` and
+  moved to this crate: 
+  - `consensus`
+  - `constants`
+- `zcash_protocol::value::Amount::into_u64`
+- `impl TryFrom<u64> for zcash_protocol::value::NonNegativeAmount`
+
+### Moved
+- `zcash_primitives::transcation::components::amount` has been moved to
+  `zcash_protocol::value`
+